Refactor: 서비스 계층 분리 - getOrThrow 메서드 생성
