+++ backend/src/main/java/com/together/backend/domain/user/service/MainPageService.java
@@ -28,16 +28,19 @@ public class MainPageService {
     private final CoupleRepository coupleRepository;
     private final UserPillRepository userPillRepository;
 
+    private User getOrThrow(String email) {
+        return userRepository.findByEmail(email).orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
+    }
+
 
     // 사용자의 이메일 받고 사용자 정보를 전달함
     public UserInfoResponse getUserInfo(String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
+        User user = getOrThrow(email);
         return new UserInfoResponse(user.getEmail(), user.getNickname(), user.getProfileImageUrl());
     }
 
     public PartnerInfoResponse getPartnerInfo(String email) {
-        // 사용자 조회
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
+        User user = getOrThrow(email);
         Optional<Couple> optCouple = coupleRepository.findByUser(user);
 
         if (optCouple.isPresent()) {
@@ -58,11 +61,10 @@ public class MainPageService {
             return new PartnerInfoResponse(null, false, 0L);
         }
 
-
     }
 
     public PillInfoResponse getPillInfo(String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
+        User user = getOrThrow(email);
         Optional<UserPill> optPill = userPillRepository.findTopByUserOrderByStartDateDesc(user);
 
         long daysOnPill = optPill.map(pill -> ChronoUnit.DAYS.between(pill.getStartDate(), LocalDate.now())).orElse(0L);
